move moddb services to moddb folder
