+++ Stalker.Gamma/ModDb/Models/ModDbRecord.cs
@@ -1,38 +1,43 @@
 using System.Security.Cryptography;
 using Stalker.Gamma.GammaInstallerServices;
+using Stalker.Gamma.Models;
 using Stalker.Gamma.Utilities;
+using ModDbService = Stalker.Gamma.ModDb.Services.ModDbService;
 
-namespace Stalker.Gamma.Models;
+namespace Stalker.Gamma.ModDb.Models;
 
-public class ModDbRecordGetMetadata(
-    string name,
-    string startLink,
-    List<string> instructions,
-    string outputDirName,
-    string gammaDir,
-    ArchiveUtility archiveUtility,
+public class ModDbRecord(
     GammaProgress gammaProgress,
-    ModDbUtility modDbUtility,
-    GetCanonicalLinkFromModDbStartLink getCanonicalLinkFromModDbStartLink,
-    GetModDbAddonMetadata getModDbAddonMetadata
+    string name,
+    string url,
+    string niceUrl,
+    string archiveName,
+    string? md5,
+    string gammaDir,
+    string outputDirName,
+    IList<string> instructions,
+    ArchiveUtility archiveUtility,
+    ModDbService modDbService
 ) : IDownloadableRecord
 {
+    private readonly GammaProgress _gammaProgress = gammaProgress;
+    private readonly string _gammaDir = gammaDir;
+    private readonly string _outputDirName = outputDirName;
+    private readonly ArchiveUtility _archiveUtility = archiveUtility;
+    private readonly ModDbService _modDbService = modDbService;
     public string Name { get; } = name;
-    public string ArchiveName { get; set; } = null!;
+    private string Url { get; } = url;
+    private string NiceUrl { get; } = niceUrl;
+    public string ArchiveName { get; } = archiveName;
+    private string? Md5 { get; } = md5;
     public string DownloadPath => Path.Join(_gammaDir, "downloads", ArchiveName);
-    public string ExtractPath => Path.Join(_gammaDir, "mods", OutputDirName);
-    public string StartLink { get; set; } = startLink;
-    public string NiceUrl { get; set; } = null!;
-    private List<string> Instructions { get; } = instructions;
-    private string OutputDirName { get; } = outputDirName;
-    public string Url => StartLink;
-    public string Md5 { get; set; } = null!;
+    private string ExtractPath => Path.Join(_gammaDir, "mods", _outputDirName);
+    private IList<string> Instructions { get; } = instructions;
 
-    public async Task DownloadAsync(CancellationToken cancellationToken)
+    public virtual async Task DownloadAsync(CancellationToken cancellationToken = default)
     {
         try
         {
-            await GetModDbAddonMetadataAsync(cancellationToken);
             if (
                 Path.Exists(DownloadPath)
                     && !string.IsNullOrWhiteSpace(Md5)
@@ -45,7 +50,7 @@ public class ModDbRecordGetMetadata(
                 || !Path.Exists(DownloadPath)
             )
             {
-                await _modDbUtility.GetModDbLinkCurl(
+                await _modDbService.GetModDbLinkCurl(
                     Url,
                     DownloadPath,
                     pct => OnProgress(GammaProgressType.Download, pct),
@@ -67,7 +72,7 @@ public class ModDbRecordGetMetadata(
         }
     }
 
-    public async Task ExtractAsync(CancellationToken cancellationToken)
+    public virtual async Task ExtractAsync(CancellationToken cancellationToken = default)
     {
         try
         {
@@ -106,19 +111,20 @@ public class ModDbRecordGetMetadata(
         }
     }
 
-    private async Task GetModDbAddonMetadataAsync(CancellationToken cancellationToken)
-    {
-        var canonicalLink = await _getCanonicalLinkFromModDbStartLink.GetCanonicalLinkAsync(
-            StartLink,
-            cancellationToken
-        );
-        var metadata = await _getModDbAddonMetadata.GetAsync(canonicalLink, cancellationToken);
-        ArchiveName = metadata.Filename;
-        Md5 = metadata.Md5Hash;
-        NiceUrl = canonicalLink;
-    }
-
     public bool Downloaded { get; set; }
+
+    public override string ToString() =>
+        $"""
+            Name: {Name}
+            Archive Name: {ArchiveName}
+            Url: {Url}
+            NiceUrl: {NiceUrl}
+            Download Path: {DownloadPath}
+            Extract Path: {ExtractPath}
+            Md5: {Md5}
+            Downloaded: {Downloaded}
+            Instructions: {string.Join(", ", Instructions)}
+            """;
 
     private void OnProgress(GammaProgressType operation, double pct) =>
         _gammaProgress.OnProgressChanged(ProgFunc(operation, pct));
@@ -137,12 +143,7 @@ public class ModDbRecordGetMetadata(
             DownloadPath = DownloadPath,
             ExtractPath = ExtractPath,
         };
-
-    private readonly GammaProgress _gammaProgress = gammaProgress;
-    private readonly string _gammaDir = gammaDir;
-    private readonly ArchiveUtility _archiveUtility = archiveUtility;
-    private readonly ModDbUtility _modDbUtility = modDbUtility;
-    private readonly GetCanonicalLinkFromModDbStartLink _getCanonicalLinkFromModDbStartLink =
-        getCanonicalLinkFromModDbStartLink;
-    private readonly GetModDbAddonMetadata _getModDbAddonMetadata = getModDbAddonMetadata;
 }
+
+public class ModDbRecordException(string message, Exception innerException)
+    : Exception(message, innerException);
