+++ Stalker.Gamma/ModDb/Models/ModDbRecordGetMetadata.cs
@@ -1,41 +1,41 @@
 using System.Security.Cryptography;
 using Stalker.Gamma.GammaInstallerServices;
+using Stalker.Gamma.Models;
 using Stalker.Gamma.Utilities;
+using ModDbGetAddonMetadataService = Stalker.Gamma.ModDb.Services.ModDbGetAddonMetadataService;
+using ModDbService = Stalker.Gamma.ModDb.Services.ModDbService;
 
-namespace Stalker.Gamma.Models;
+namespace Stalker.Gamma.ModDb.Models;
 
-public class ModDbRecord(
-    GammaProgress gammaProgress,
+public class ModDbRecordGetMetadata(
     string name,
-    string url,
-    string niceUrl,
-    string archiveName,
-    string? md5,
-    string gammaDir,
+    string startLink,
+    List<string> instructions,
     string outputDirName,
-    IList<string> instructions,
+    string gammaDir,
     ArchiveUtility archiveUtility,
-    ModDbUtility modDbUtility
+    GammaProgress gammaProgress,
+    ModDbService modDbService,
+    GetCanonicalLinkFromModDbStartLink getCanonicalLinkFromModDbStartLink,
+    ModDbGetAddonMetadataService modDbGetAddonMetadataService
 ) : IDownloadableRecord
 {
-    private readonly GammaProgress _gammaProgress = gammaProgress;
-    private readonly string _gammaDir = gammaDir;
-    private readonly string _outputDirName = outputDirName;
-    private readonly ArchiveUtility _archiveUtility = archiveUtility;
-    private readonly ModDbUtility _modDbUtility = modDbUtility;
     public string Name { get; } = name;
-    private string Url { get; } = url;
-    private string NiceUrl { get; } = niceUrl;
-    public string ArchiveName { get; } = archiveName;
-    private string? Md5 { get; } = md5;
+    public string ArchiveName { get; set; } = null!;
     public string DownloadPath => Path.Join(_gammaDir, "downloads", ArchiveName);
-    private string ExtractPath => Path.Join(_gammaDir, "mods", _outputDirName);
-    private IList<string> Instructions { get; } = instructions;
+    public string ExtractPath => Path.Join(_gammaDir, "mods", OutputDirName);
+    public string StartLink { get; set; } = startLink;
+    public string NiceUrl { get; set; } = null!;
+    private List<string> Instructions { get; } = instructions;
+    private string OutputDirName { get; } = outputDirName;
+    public string Url => StartLink;
+    public string Md5 { get; set; } = null!;
 
-    public virtual async Task DownloadAsync(CancellationToken cancellationToken = default)
+    public async Task DownloadAsync(CancellationToken cancellationToken)
     {
         try
         {
+            await GetModDbAddonMetadataAsync(cancellationToken);
             if (
                 Path.Exists(DownloadPath)
                     && !string.IsNullOrWhiteSpace(Md5)
@@ -48,7 +48,7 @@ public class ModDbRecord(
                 || !Path.Exists(DownloadPath)
             )
             {
-                await _modDbUtility.GetModDbLinkCurl(
+                await _modDbService.GetModDbLinkCurl(
                     Url,
                     DownloadPath,
                     pct => OnProgress(GammaProgressType.Download, pct),
@@ -70,7 +70,7 @@ public class ModDbRecord(
         }
     }
 
-    public virtual async Task ExtractAsync(CancellationToken cancellationToken = default)
+    public async Task ExtractAsync(CancellationToken cancellationToken)
     {
         try
         {
@@ -109,20 +109,22 @@ public class ModDbRecord(
         }
     }
 
-    public bool Downloaded { get; set; }
+    private async Task GetModDbAddonMetadataAsync(CancellationToken cancellationToken)
+    {
+        var canonicalLink = await _getCanonicalLinkFromModDbStartLink.GetCanonicalLinkAsync(
+            StartLink,
+            cancellationToken
+        );
+        var metadata = await _modDbGetAddonMetadataService.GetAsync(
+            canonicalLink,
+            cancellationToken
+        );
+        ArchiveName = metadata.Filename;
+        Md5 = metadata.Md5Hash;
+        NiceUrl = canonicalLink;
+    }
 
-    public override string ToString() =>
-        $"""
-            Name: {Name}
-            Archive Name: {ArchiveName}
-            Url: {Url}
-            NiceUrl: {NiceUrl}
-            Download Path: {DownloadPath}
-            Extract Path: {ExtractPath}
-            Md5: {Md5}
-            Downloaded: {Downloaded}
-            Instructions: {string.Join(", ", Instructions)}
-            """;
+    public bool Downloaded { get; set; }
 
     private void OnProgress(GammaProgressType operation, double pct) =>
         _gammaProgress.OnProgressChanged(ProgFunc(operation, pct));
@@ -141,7 +143,13 @@ public class ModDbRecord(
             DownloadPath = DownloadPath,
             ExtractPath = ExtractPath,
         };
-}
 
-public class ModDbRecordException(string message, Exception innerException)
-    : Exception(message, innerException);
+    private readonly GammaProgress _gammaProgress = gammaProgress;
+    private readonly string _gammaDir = gammaDir;
+    private readonly ArchiveUtility _archiveUtility = archiveUtility;
+    private readonly ModDbService _modDbService = modDbService;
+    private readonly GetCanonicalLinkFromModDbStartLink _getCanonicalLinkFromModDbStartLink =
+        getCanonicalLinkFromModDbStartLink;
+    private readonly ModDbGetAddonMetadataService _modDbGetAddonMetadataService =
+        modDbGetAddonMetadataService;
+}
